+++ src/tiled/automapper.h
@@ -206,7 +206,7 @@ struct TILED_EDITOR_EXPORT AutoMappingContext
     std::vector<std::unique_ptr<Layer>> newLayers;  // Layers created in AutoMapper::prepareAutoMap
     QVector<QVector<AddMapObjects::Entry>> newMapObjects;   // Objects placed by AutoMapper
     QSet<MapObject*> mapObjectsToRemove;
-    QHash<Layer*, Properties> changedProperties;
+    QHash<const Layer*, Properties> changedProperties;
 
     // Clones of existing tile layers that might have been changed in AutoMapper::autoMap
     std::unordered_map<TileLayer*, std::unique_ptr<TileLayer>> originalToOutputLayerMapping;